Finals Filipino Family Income and Expenditure Act
The dataset comes from the Family Income and Expenditure Survey (FIES) conducted by the Philippine Statistics Authority (PSA). This national survey is carried out every three years to collect information on family income and expenditures, including levels of consumption by item of expenditure, sources of income in cash, and other related data that influence income and spending patterns in the Philippines.

The dataset contains more than 40,000 observations and 60 variables, primarily comprising the household income and expenditures of specific households. For this analysis, we will be using Python’s pandas, numpy, and folium libraries.